requirements.txt
Flask
numpy
matplotlib
plotly
gnpy
python-dotenv
